Move the conversion functions into appropriate places
This stops structured_datat growing if more formats are added
